tests: Make ToStringTests partial of QuantityTests
